fix(FollowStatsVO): UserRepository followStats 조회 시 오류 수정 #18

+++ goaltracker/src/main/java/com/goaltracker/user/dto/UserProfileWithFollowStatsDTO.java
@@ -10,8 +10,8 @@ import java.util.List;
 @Builder
 public class UserProfileWithFollowStatsDTO {
     String username;
-    int followerCount;
-    int followingCount;
+    long followerCount;
+    long followingCount;
     String description;
     List<String> interests;
     RelationType relationType;
