+++ goaltracker/src/main/java/com/goaltracker/user/dto/vo/FollowStatsVO.java
@@ -4,10 +4,10 @@ import lombok.Getter;
 
 @Getter
 public class FollowStatsVO {
-    int followerCount;
-    int followingCount;
+    long followerCount;
+    long followingCount;
 
-    public FollowStatsVO(int followerCount, int followingCount) {
+    public FollowStatsVO(long followerCount, long followingCount) {
         this.followerCount = followerCount;
         this.followingCount = followingCount;
     }
